test: Stabilize playback handoff coverage
Split oversized playback tests and cover the new-mix handoff at the provider boundary.
